backend/restaurante/requirements.txt
apparmor==4.0.3
application-utility==1.3.2
argcomplete==3.4.0
asgiref==3.8.1
attrs==23.2.1.dev0
beautifulsoup4==4.12.3
btrfsutil==6.10
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
configobj==5.0.8
cryptography==43.0.3
cupshelpers==1.0
dbus-python==1.3.2
defusedxml==0.7.1
distro==1.9.0
Django==5.1.2
djangorestframework==3.15.2
duplicity==3.0.3
fasteners==0.19
filelock==3.16.1
gufw==24.4.0
idna==3.10
layoutswitcherlib==0.8.38
LibAppArmor==4.0.3
libfdt==1.7.2
mysqlclient==2.2.4
netsnmp-python==1.0a1
nftables==0.1
npyscreen==4.10.5
packaging==24.2
pacman_mirrors==4.27
paho-mqtt==1.6.1
pika==1.3.2
pillow==11.0.0
pipx==1.7.1
platformdirs==4.3.6
pwquality==1.4.5
pycairo==1.27.0
pycparser==2.22
pycups==2.0.4
PyGObject==3.50.0
pytz==2024.2
requests==2.32.3
requests-file==1.5.1
setproctitle==1.3.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
tldextract==5.1.3
ufw==0.36.2
urllib3==1.26.20
userpath==1.9.2
validate==5.0.8
wheel==0.45.0
